refactor(task-action): extract the diagnostics collaborator (S2)
TaskAction was a 1212-line God class mixing two concerns: action
classification (Task+Marker → key/label/command) and red-status
diagnostics (the summary/detail/artifact_paths/suggested_next_action
payload). This splits the diagnostics concern into a focused PORO,
Hive::TaskAction::Diagnostic (lib/hive/task_action/diagnostic.rb).

The new collaborator is a pure function of (task, marker) plus the
classification context it depends on — the action key, the rerun
command, and the five state predicates that also drive classification
(incomplete_plan_artifact?, finalize_missing_pr_md?/_pr_url?/_url_mismatch?,
legacy_execute_findings?) plus the stale-agent reason. TaskAction
computes those once and passes the results in, so the classifier and the
diagnostic can never disagree about a (task, marker) pair, and Diagnostic
never reaches back into the classifier's private surface.

Moved to Diagnostic: diagnostic/next_action payload assembly,
synthetic_stale_agent_diagnostic, diagnostic_action?, every *_artifacts
builder, the log/path/frontmatter helpers, marker_summary/marker_detail,
suggested_next_action_payload + retry_command_string, the redact/truncate
helpers, and the DIAGNOSTIC_*/ARTIFACT_PATHS_MAX/LOG_GLOB_CAP/
FRONTMATTER_SCAN_BYTES/AUTO_COMMIT_*/FIX_STATUS_CHECK_* constants.

Stayed in TaskAction: action classification + command emission, the
shared state predicates and log_dirs (classification inputs), the
command_prefix/stage_dir builders, and the two public class methods
(self.marker_signature, self.max_passes_review_stale_with_escalations?)
that external consumers call. TaskAction#diagnostic/#next_action now
delegate to a memoized Diagnostic; #payload's shape is unchanged.

The hive status --json diagnostic payload is byte-identical for every
marker state (status/daemon golden tests guard it). De-noised the moved
methods with the same DHH/Kieran pass as S1 (endless predicate defs,
or-raise reads, stripped what-comments).

task_action.rb: 1212 → 634 lines; new diagnostic.rb: 646 lines.
